requirements.txt
kagglehub
apache-airflow-providers-snowflake 
snowflake-connector-python
